+++ nodes.py
@@ -8,11 +8,9 @@ import hashlib
 import traceback
 import math
 import time
-import random
 import logging
 
-from PIL import Image, ImageOps, ImageSequence
-from PIL.PngImagePlugin import PngInfo
+from PIL import Image, ImageOps
 
 import numpy as np
 import safetensors.torch
@@ -1551,181 +1549,6 @@ class KSamplerAdvanced:
         return common_ksampler(model, noise_seed, steps, cfg, sampler_name, scheduler, positive, negative, latent_image, denoise=denoise, disable_noise=disable_noise, start_step=start_at_step, last_step=end_at_step, force_full_denoise=force_full_denoise)
 
 
-class SaveImage(io.ComfyNodeV3):
-    @classmethod
-    def DEFINE_SCHEMA(cls):
-        return io.SchemaV3(
-            node_id="SaveImage",
-            display_name="Save Image",
-            description="Saves the input images to your ComfyUI output directory.",
-            category="image",
-            inputs=[
-                io.Image.Input(
-                    "images",
-                    display_name="images",
-                    tooltip="The images to save.",
-                ),
-                io.String.Input(
-                    "filename_prefix",
-                    default="ComfyUI",
-                    tooltip="The prefix for the file to save. This may include formatting information such as %date:yyyy-MM-dd% or %Empty Latent Image.width% to include values from nodes.",
-                ),
-            ],
-            hidden=[io.Hidden.prompt, io.Hidden.extra_pnginfo],
-            is_output_node=True,
-        )
-
-    def __init__(self):
-        super().__init__()
-        self.output_dir = folder_paths.get_output_directory()
-        self.type = "output"
-        self.prefix_append = ""
-        self.compress_level = 4
-
-    def execute(self, images, filename_prefix="ComfyUI", prompt=None, extra_pnginfo=None):
-        filename_prefix += self.prefix_append
-        full_output_folder, filename, counter, subfolder, filename_prefix = folder_paths.get_save_image_path(filename_prefix, self.output_dir, images[0].shape[1], images[0].shape[0])
-        results = list()
-        for (batch_number, image) in enumerate(images):
-            i = 255. * image.cpu().numpy()
-            img = Image.fromarray(np.clip(i, 0, 255).astype(np.uint8))
-            metadata = None
-            if not args.disable_metadata:
-                metadata = PngInfo()
-                if prompt is not None:
-                    metadata.add_text("prompt", json.dumps(prompt))
-                if extra_pnginfo is not None:
-                    for x in extra_pnginfo:
-                        metadata.add_text(x, json.dumps(extra_pnginfo[x]))
-
-            filename_with_batch_num = filename.replace("%batch_num%", str(batch_number))
-            file = f"{filename_with_batch_num}_{counter:05}_.png"
-            img.save(os.path.join(full_output_folder, file), pnginfo=metadata, compress_level=self.compress_level)
-            results.append({
-                "filename": file,
-                "subfolder": subfolder,
-                "type": self.type,
-            })
-            counter += 1
-
-        return { "ui": { "images": results } }
-
-
-class PreviewImage(SaveImage):
-    @classmethod
-    def DEFINE_SCHEMA(cls):
-        return io.SchemaV3(
-            node_id="PreviewImage",
-            display_name="Preview Image",
-            description="Preview the input images.",
-            category="image",
-            inputs=[
-                io.Image.Input(
-                    "images",
-                    display_name="images",
-                    tooltip="The images to preview.",
-                ),
-            ],
-            hidden=[io.Hidden.prompt, io.Hidden.extra_pnginfo],
-            is_output_node=True,
-        )
-
-    def __init__(self):
-        super().__init__()
-        self.output_dir = folder_paths.get_temp_directory()
-        self.type = "temp"
-        self.prefix_append = "_temp_" + ''.join(random.choice("abcdefghijklmnopqrstupvxyz") for x in range(5))
-        self.compress_level = 1
-
-
-class LoadImage(io.ComfyNodeV3):
-    @classmethod
-    def DEFINE_SCHEMA(cls):
-        return io.SchemaV3(
-            node_id="LoadImage",
-            display_name="Load Image",
-            category="image",
-            inputs=[
-                io.Combo.Input(
-                    "image",
-                    display_name="image",
-                    image_upload=True,
-                    image_folder=io.FolderType.input,
-                    content_types=["image"],
-                ),
-            ],
-            outputs=[
-                io.Image.Output(
-                    "IMAGE",
-                ),
-                io.Mask.Output(
-                    "MASK",
-                ),
-            ],
-        )
-
-    @classmethod
-    def execute(cls, image) -> io.NodeOutput:
-        img = node_helpers.pillow(Image.open, folder_paths.get_annotated_filepath(image))
-
-        output_images = []
-        output_masks = []
-        w, h = None, None
-
-        excluded_formats = ['MPO']
-
-        for i in ImageSequence.Iterator(img):
-            i = node_helpers.pillow(ImageOps.exif_transpose, i)
-
-            if i.mode == 'I':
-                i = i.point(lambda i: i * (1 / 255))
-            image = i.convert("RGB")
-
-            if len(output_images) == 0:
-                w = image.size[0]
-                h = image.size[1]
-
-            if image.size[0] != w or image.size[1] != h:
-                continue
-
-            image = np.array(image).astype(np.float32) / 255.0
-            image = torch.from_numpy(image)[None,]
-            if 'A' in i.getbands():
-                mask = np.array(i.getchannel('A')).astype(np.float32) / 255.0
-                mask = 1. - torch.from_numpy(mask)
-            elif i.mode == 'P' and 'transparency' in i.info:
-                mask = np.array(i.convert('RGBA').getchannel('A')).astype(np.float32) / 255.0
-                mask = 1. - torch.from_numpy(mask)
-            else:
-                mask = torch.zeros((64,64), dtype=torch.float32, device="cpu")
-            output_images.append(image)
-            output_masks.append(mask.unsqueeze(0))
-
-        if len(output_images) > 1 and img.format not in excluded_formats:
-            output_image = torch.cat(output_images, dim=0)
-            output_mask = torch.cat(output_masks, dim=0)
-        else:
-            output_image = output_images[0]
-            output_mask = output_masks[0]
-
-        return io.NodeOutput(output_image, output_mask)
-
-    @classmethod
-    def IS_CHANGED(s, image):
-        image_path = folder_paths.get_annotated_filepath(image)
-        m = hashlib.sha256()
-        with open(image_path, 'rb') as f:
-            m.update(f.read())
-        return m.digest().hex()
-
-    @classmethod
-    def VALIDATE_INPUTS(s, image):
-        if not folder_paths.exists_annotated_filepath(image):
-            return "Invalid image file: {}".format(image)
-
-        return True
-
-
 class LoadImageMask:
     _color_channels = ["alpha", "red", "green", "blue"]
     @classmethod
@@ -1774,28 +1597,6 @@ class LoadImageMask:
             return "Invalid image file: {}".format(image)
 
         return True
-
-
-class LoadImageOutput(LoadImage):
-    @classmethod
-    def INPUT_TYPES(s):
-        return {
-            "required": {
-                "image": ("COMBO", {
-                    "image_upload": True,
-                    "image_folder": "output",
-                    "remote": {
-                        "route": "/internal/files/output",
-                        "refresh_button": True,
-                        "control_after_refresh": "first",
-                    },
-                }),
-            }
-        }
-
-    DESCRIPTION = "Load an image from the output folder. When the refresh button is clicked, the node will update the image list and automatically select the first image, allowing for easy iteration."
-    EXPERIMENTAL = True
-    FUNCTION = "load_image"
 
 
 class ImageScale:
@@ -1980,11 +1781,7 @@ NODE_CLASS_MAPPINGS = {
     "LatentUpscaleBy": LatentUpscaleBy,
     "LatentFromBatch": LatentFromBatch,
     "RepeatLatentBatch": RepeatLatentBatch,
-    "SaveImage": SaveImage,
-    "PreviewImage": PreviewImage,
-    "LoadImage": LoadImage,
     "LoadImageMask": LoadImageMask,
-    "LoadImageOutput": LoadImageOutput,
     "ImageScale": ImageScale,
     "ImageScaleBy": ImageScaleBy,
     "ImageInvert": ImageInvert,
@@ -2081,11 +1878,7 @@ NODE_DISPLAY_NAME_MAPPINGS = {
     "LatentFromBatch" : "Latent From Batch",
     "RepeatLatentBatch": "Repeat Latent Batch",
     # Image
-    "SaveImage": "Save Image",
-    "PreviewImage": "Preview Image",
-    "LoadImage": "Load Image",
     "LoadImageMask": "Load Image (as Mask)",
-    "LoadImageOutput": "Load Image (from Outputs)",
     "ImageScale": "Upscale Image",
     "ImageScaleBy": "Upscale Image By",
     "ImageUpscaleWithModel": "Upscale Image (using Model)",
@@ -2295,7 +2088,6 @@ def init_builtin_extra_nodes():
         "nodes_align_your_steps.py",
         "nodes_attention_multiply.py",
         "nodes_advanced_samplers.py",
-        "nodes_webcam.py",
         "nodes_audio.py",
         "nodes_sd3.py",
         "nodes_gits.py",
@@ -2330,6 +2122,9 @@ def init_builtin_extra_nodes():
         "nodes_tcfg.py"
         "nodes_v3_test.py",
         "nodes_v1_test.py",
+        "v3/nodes_images.py",
+        "v3/nodes_mask.py",
+        "v3/nodes_webcam.py",
     ]
 
     import_failed = []
